requirements.txt
python-dotenv==0.21.1
watchdog==3.0.0
pyftpdlib==2.1.0
